Fix staticcheck linter warnings
By moving darwin specific codes to darwin file.
