+++ staticdns.go
@@ -1,12 +1,8 @@
 package ctrld
 
 import (
-	"bufio"
-	"bytes"
-	"io"
 	"net"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"runtime"
 	"strings"
@@ -80,39 +76,4 @@ func SavedStaticNameservers(iface *net.Interface) ([]string, string) {
 		ns = append(ns, v)
 	}
 	return ns, file
-}
-
-func patchNetIfaceName(iface *net.Interface) (bool, error) {
-	b, err := exec.Command("networksetup", "-listnetworkserviceorder").Output()
-	if err != nil {
-		return false, err
-	}
-
-	patched := false
-	if name := networkServiceName(iface.Name, bytes.NewReader(b)); name != "" {
-		patched = true
-		iface.Name = name
-	}
-	return patched, nil
-}
-
-func networkServiceName(ifaceName string, r io.Reader) string {
-	scanner := bufio.NewScanner(r)
-	prevLine := ""
-	for scanner.Scan() {
-		line := scanner.Text()
-		if strings.Contains(line, "*") {
-			// Network services is disabled.
-			continue
-		}
-		if !strings.Contains(line, "Device: "+ifaceName) {
-			prevLine = line
-			continue
-		}
-		parts := strings.SplitN(prevLine, " ", 2)
-		if len(parts) == 2 {
-			return strings.TrimSpace(parts[1])
-		}
-	}
-	return ""
 }
